requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
passlib==1.7.4
pwdlib==0.2.1
PyJWT==2.9.0
pydantic-settings==2.6.1
python-multipart==0.0.19
email-validator
pytest
pytest-mock
httpx